+++ packages/svelte/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Next
 
 * Update to Svelte v2 minimum dependency.
-* Remove UMD builds.
+* Remove `<script>` builds.
 
 ## 1.0.0-beta.5